IngestIndexerService/requirements.txt
aiokafka==0.10.0
fastapi==0.115.0  # не используется, но для консистентности
elasticsearch[async]==8.15.1
httpx==0.27.2
pymupdf==1.24.10
python-docx==1.1.2
pytesseract==0.3.13
pillow==10.4.0
langchain-text-splitters==0.3.0
structlog==24.4.0